Created the current projects section
Got the project data displayed on the current projects section. Does not have click functionality yet and needs to be visually altered. Also need to work on prev projects section, but want to fill out the projects table in data.js beforehand so I have data to look at to see it visually working